requirements.txt
google-api-python-client
python-dotenv
google-generativeai
load_dotenv

